+++ NEWS.md
@@ -1,3 +1,8 @@
+forplot 3.0.1
+------------------
+add option to drop individual headers and footers in combinefobj
+add padding to xlims
+
 forplot 3.0.0
 ------------------
 add combinefobj to combine fobj